refactor: update import paths for user-related components and remove obsolete account pages
